requirements.txt
Flask
prometheus_client
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-jaeger
opentelemetry-exporter-prometheus
opentelemetry-instrumentation-flask
deprecated
python-logstash

